Fix release build
__Motivation__

Currently the release process runs two Gradle commands, with only one of them having the `releaseBuild` flag set.

This enables the logic that automatically adds "-SNAPSHOT" to kick in, leading to resource files generated with the wrong version number.

__Modifications__

- Set the `releaseBuild` flag for all Gradle commands used at release time.
- Improve the logic that auto adds "-SNAPSHOT" so it logs when it kicks in and also applies to all sub-projects.

__Results__

Release builds do not contain snasphot versions.
