+++ docker/docker-compose.yaml
@@ -80,7 +80,9 @@ services:
       - CI
       - BINTRAY_USER
       - BINTRAY_KEY
-    command: bash -cl "./gradlew --no-daemon clean check && ./gradlew --no-daemon bintrayUpload -PreleaseBuild=true"
+    command: >
+      bash -cl "./gradlew --no-daemon -PreleaseBuild=true clean check &&
+      ./gradlew --no-daemon -PreleaseBuild=true bintrayUpload"
 
   shell:
     << : *common
